+++ src/components/feature/request/CreateRequestForm.css
@@ -99,6 +99,7 @@
   font-weight: bold;
   color: #333;
   margin-bottom: 10px;
+  text-align: left;
 }
 
 .formSection {
